Fix path calculation for ODS source file
The original.ods file was using "$CWD/original.ods" which is not valid in
the case where you run the script from another directory (i.e.
python ezfio/ezfio.py ...).

Use realpath() of the script file instead, and check that the file exists
at this spot before even starting a test to ensure runs don't fail after many
hours.

+++ ezfio.py
@@ -314,7 +314,10 @@ def SetupFiles():
     AppendFile("IOPS", timeseriescsv) # Add IOPS header
 
     # ODS input and output files
-    odssrc = os.getcwd() + "/original.ods"
+    odssrc = os.path.dirname( os.path.realpath(__file__) + "/original.ods" )
+    if not os.path.exists(odssrc):
+        print "ERROR: Can't find original ODS spreadsheet '" + odssrc + "'. ",
+        sys.exit(1)
     odsdest = outputDest + "/ezfio_results_"+suffix+".ods"
     if os.path.exists(odsdest):
         os.unlink(odsdest)
